# Dokumentáció/Fejlesztői napló/Api/pipeline_swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PipeLine.Backend",
    "version": "1.0"
  },
  "paths": {
...
    "/api/ChargingInstance/myFilteredInstances": {
      "get": {
        "tags": [
          "ChargingInstance"
        ],
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ChargingInstance/myFilteredFinishedInstances": {
      "get": {
        "tags": [
          "ChargingInstance"
        ],
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/ChargingInstance": {
      "post": {
        "tags": [
          "ChargingInstance"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChargingInstanceDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChargingInstanceDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChargingInstanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "ChargingInstance"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "ChargingInstance"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChargingInstanceDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChargingInstanceDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChargingInstanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ChargingInstance/finishInstance": {
      "put": {
        "tags": [
          "ChargingInstance"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChargingInstanceDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChargingInstanceDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChargingInstanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ChargingInstance/ChargingProgress/{instanceId}": {
      "put": {
        "tags": [
          "ChargingInstance"
        ],
        "parameters": [
          {
            "name": "instanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ChargingInstance/{id}": {
      "get": {
        "tags": [
          "ChargingInstance"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "ChargingInstance"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/ChargingPort/FilteredPortsOfChargingStation/{id}": {
      "get": {
        "tags": [
          "ChargingPort"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "input",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ChargingPort/NumberOfPortsOfChargingStation/{id}": {
      "get": {
        "tags": [
          "ChargingPort"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ChargingPort": {
      "get": {
        "tags": [
          "ChargingPort"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "ChargingPort"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChargingPortDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChargingPortDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChargingPortDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "ChargingPort"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChargingPortDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChargingPortDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChargingPortDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ChargingPort/{id}": {
      "get": {
        "tags": [
          "ChargingPort"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "ChargingPort"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChargingInstanceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "chargingPortId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "deviceId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "startingPercentage": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "endPercentage": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "desiredEndPercentage": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "portNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "chargingStationName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ChargingPortDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "isBeingUsed": {
            "type": "boolean"
          },
          "isDisabled": {
            "type": "boolean"
          },
          "chargingStationId": {
            "type": "string",
            "format": "uuid"
          },
          "portNumber": {
            "type": "integer",
            "format": "int32"
          },
          "maxChargingSpeed": {
            "type": "number",
            "format": "double"
          },
          "isCharging": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}